🚚  data불러오는 기능들 이동
